requirements.txt
pydantic
fpdf2
selenium
webdriver-manager
openpyxl
pyinstaller
PySide6
pyyaml
requests